refactor(core): split public types, scale collect/train, and candidates
Extract types-public into advisory/model/options modules; scale-axis
evidence into x/y collectors; pipeline scale training into position vs
color; boxplot body into layout vs batch assembly; and candidate logical
value resolution into a pure helper. Facades keep import paths stable.
Add characterization tests for resolveCandidateLogicalValues.
